ui: refactor settingsview to settings directory - split `SecuritySettingsView` into its own file.
